pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-async-manager"
version = "0.1.0"
description = "Django library for managing asynchronous tasks with scheduling and dependency management"
readme = "README.md"
authors = [
    {name = "Michał Konwiak", email = "michalkonwiak1@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.12"
dependencies = [
    "croniter>=6.0.0",
    "django>=3.2",
    "setuptools>=78.1.0",
]

[project.optional-dependencies]
dev = [
    "coverage>=7.8.0",
    "django-stubs>=5.1.3",
    "factory-boy>=3.3.3",
    "flake8>=7.1.2",
    "pre-commit>=4.1.0",
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "ruff>=0.9.10",
]

[project.urls]
Homepage = "https://github.com/michalkonwiak/django-async-manager"
Documentation = "https://github.com/michalkonwiak/django-async-manager"
Repository = "https://github.com/michalkonwiak/django-async-manager.git"
Issues = "https://github.com/michalkonwiak/django-async-manager/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["tests*"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.rst", "*.json", "*.html", "*.css", "*.js", "*.png", "*.jpg", "*.jpeg", "*.gif", "*.svg"]

[ruff]
line-length = 100
target-version = "py312"

exclude = [
    ".git",
    "__pycache__",
    "migrations",
    "venv",
    ".venv",
    "env",
    "settings.py",
    "manage.py"
]

ignore = [
    "E203", "W503", "E501", "F401", "F811", "B008"
]

[ruff.mccabe]
max-complexity = 10
